refactor: small eusart code cleanup
