+++ firmware/development/kampela_experiments_efm32pg23/app3/src/devices/mod.rs
@@ -1,4 +1,5 @@
 pub mod power;
+pub mod psram;
 pub mod display;
 pub mod se_rng;
 pub mod se_aes_gcm;
