+++ firmware/development/kampela_experiments_efm32pg23/app3/src/peripherals/eusart.rs
@@ -129,8 +129,7 @@ pub fn init_eusart(peripherals: &mut Peripherals) {
     while peripherals.EUSART2_S.status.read().rxidle().bit_is_clear()
         | peripherals.EUSART2_S.status.read().txidle().bit_is_clear()  {}
 
-    psram_reset(peripherals); // reset here, right after setup
-
+    // remember to reset connected ram device here later, right after setup
 }
 
 fn eusart_disable(peripherals: &mut Peripherals) {
@@ -220,36 +219,4 @@ fn eusart_reset(peripherals: &mut Peripherals) {
     peripherals.EUSART2_S.clkdiv.reset();
 }
 
-pub fn psram_reset(peripherals: &mut Peripherals) {
-    psram_chip_select_set(peripherals); // deselect PSRAM
-    psram_chip_select_clear(peripherals); // select PSRAM, explain why
-    psram_write_read_byte(peripherals, PSRAM_RESET_ENABLE);
-    psram_chip_select_set(peripherals); // deselect PSRAM
-    psram_chip_select_clear(peripherals); // select PSRAM
-    psram_write_read_byte(peripherals, PSRAM_RESET);
-    psram_chip_select_set(peripherals); // deselect PSRAM
-}
-
-pub fn psram_write_read_byte(peripherals: &mut Peripherals, byte: u8) -> u8 {
-    while peripherals.EUSART2_S.status.read().txfl().bit_is_clear() {}
-    peripherals.EUSART2_S.txdata.write({|w_reg|
-        w_reg
-            // EUSART tx and rx are u16,
-            // single byte is used here because of the commands,
-            // setting used is `.databits().eight()`
-            .txdata().variant(byte as u16)
-    });
-    while peripherals.EUSART2_S.status.read().rxfl().bit_is_clear() {}
-    peripherals.EUSART2_S.rxdata.read().rxdata().bits().try_into().expect("configured frame for 8 data bits")
-}
-
-/// PSRAM commands from manual
-pub const PSRAM_RESET_ENABLE: u8 = 0x66;
-pub const PSRAM_RESET: u8 = 0x99;
-pub const PSRAM_READ_ID: u8 = 0x9f;
-pub const PSRAM_READ: u8 = 0x03;
-pub const PSRAM_WRITE: u8 = 0x02;
-
-pub const ID_LEN: usize = 3;
-pub const ADDR_LEN: usize = 3;
